# repository-server/src/main/resources/openapi/graph-repository.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Graph Atlas - Administrative API",
    "version": "0.5.0-preview",
    "description": "Spring Boot 4.1.1. JSON body schemas intentionally remain extensible for format-specific repository settings. See docs/en/api.md for exact fields and docs/en/coverage.md for protocol scope. Administrative error messages support en, de and fa through Accept-Language; English is the default. Machine error codes and package identifiers are not translated."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "security": [
    {
      "basicAuth": []
    },
    {
      "bearerAuth": []
    },
    {
      "sessionAuth": []
    }
  ],
  "components": {
    "securitySchemes": {
      "basicAuth": {
        "type": "http",
        "scheme": "basic"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "opaque gr_ token"
      },
      "sessionAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "gr_session"
      }
    },
    "schemas": {
      "AtlasAssetReference": {
        "type": "object",
        "required": [
          "repo",
          "path"
        ],
        "properties": {
          "repo": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "AtlasCapture": {
        "type": "object",
        "required": [
          "name",
          "version",
          "assets"
        ],
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 100
          },
          "version": {
            "type": "string",
            "maxLength": 100
          },
          "note": {
            "type": "string",
            "maxLength": 2000
          },
          "assets": {
            "type": "array",
            "minItems": 1,
            "items": {
              "$ref": "#/components/schemas/AtlasAssetReference"
            }
          }
        }
      },
      "AtlasDecision": {
        "type": "object",
        "required": [
          "expectedRevision",
          "note"
        ],
        "properties": {
          "expectedRevision": {
            "type": "integer",
            "minimum": 1
          },
          "note": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          }
        }
      },
      "AtlasHold": {
        "type": "object",
        "required": [
          "sha256",
          "active",
          "reason",
          "expectedRevision"
        ],
        "properties": {
          "sha256": {
            "type": "string",
            "pattern": "^[a-f0-9]{64}$"
          },
          "active": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "minLength": 1,
            "maxLength": 2000
          },
          "expectedRevision": {
            "type": "integer",
            "minimum": 0
          }
        }
      },
      "AtlasGate": {
        "type": "object",
        "required": [
          "releaseId",
          "state",
          "revision",
          "deep",
          "allowed",
          "integrityOk",
          "signatureValid",
          "problems",
          "pointInTime"
        ],
        "properties": {
          "releaseId": {
            "type": "string",
            "format": "uuid"
          },
          "state": {
            "type": "string",
            "enum": [
              "DRAFT",
              "IN_REVIEW",
              "APPROVED",
              "RELEASED",
              "REJECTED",
              "REVOKED"
            ]
          },
          "revision": {
            "type": "integer"
          },
          "deep": {
            "type": "boolean"
          },
          "allowed": {
            "type": "boolean"
          },
          "integrityOk": {
            "type": "boolean"
          },
          "signatureValid": {
            "type": "boolean"
          },
          "problems": {
            "type": "array",
            "items": {
              "type": "object"
            }
          },
          "checkedDigests": {
            "type": "integer"
          },
          "sourceDrift": {
            "type": "integer"
          },
          "checkedAt": {
            "type": "string",
            "format": "date-time"
          },
          "pointInTime": {
            "type": "boolean"
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/releases": {
      "get": {
        "tags": [
          "Release assurance"
        ],
        "summary": "List permission-scoped release capsules",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "parameters": [
          {
            "name": "state",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "",
                "DRAFT",
                "IN_REVIEW",
                "APPROVED",
                "RELEASED",
                "REJECTED",
                "REVOKED"
              ]
            }
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 100
            }
          }
        ]
      },
      "post": {
        "tags": [
          "Release assurance"
        ],
        "summary": "Create immutable snapshot; requires write and read on all sources",
        "responses": {
          "201": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtlasCapture"
              }
            }
          }
        }
      }
    },
    "/api/releases/{id}": {
      "get": {
        "tags": [
          "Release assurance"
        ],
        "summary": "Read a capsule with manifest and allowed actions",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/releases/{id}/transitions/{action}": {
      "post": {
        "tags": [
          "Release assurance"
        ],
        "summary": "Submit, approve, reject, release or revoke; optimistic revision required",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "action",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "submit",
                "approve",
                "reject",
                "release",
                "revoke"
              ]
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtlasDecision"
              }
            }
          }
        }
      }
    },
    "/api/releases/{id}/gate": {
      "get": {
        "tags": [
          "Release assurance"
        ],
        "summary": "Point-in-time deployment gate; deep SHA-256 checks by default",
        "responses": {
          "200": {
            "description": "Successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AtlasGate"
                }
              }
            }
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "deep",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": true
            }
          }
        ]
      }
    },
    "/api/releases/{id}/evidence": {
      "get": {
        "tags": [
          "Release assurance"
        ],
        "summary": "Export signed historical release evidence",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ]
      }
    },
    "/api/releases/{id}/assets/{index}": {
      "get": {
        "tags": [
          "Release assurance"
        ],
        "summary": "Download pinned content; released state and no digest hold required",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "423": {
            "description": "Digest is quarantined"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "index",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 0
            }
          }
        ]
      }
    },
    "/api/release-signing-key": {
      "get": {
        "tags": [
          "Release assurance"
        ],
        "summary": "Read instance signing fingerprint (admin)",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    },
    "/api/quarantine": {
      "get": {
        "tags": [
          "Release assurance"
        ],
        "summary": "List digest holds (admin)",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      },
      "post": {
        "tags": [
          "Release assurance"
        ],
        "summary": "Create or resolve a digest hold (admin)",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AtlasHold"
              }
            }
          }
        }
      }
    },
    "/api/insights/storage": {
      "get": {
        "tags": [
          "Release assurance"
        ],
        "summary": "Read measured storage accounting within current permissions",
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "401": {
            "description": "Authentication required"
          },
          "403": {
            "description": "Insufficient permission or self-approval"
          },
          "409": {
            "description": "Stale revision, invalid transition or integrity failure"
          },
          "400": {
            "description": "Invalid request"
          },
          "404": {
            "description": "Resource not found"
          }
        }
      }
    }
  }
}